SWYVRAE-547: USB OTG DWC2 driver: The parity fix in isochronous mode
USB OTG driver in isochronous mode has to set the parity of the receiving
microframe. The parity is set to even by default. This causes problems for
an audio gadget, if the host starts transmitting on odd microframes.

This fix uses Incomplete Periodic Transfer interrupt to toggle between
even and odd parity until the Transfer Complete interrupt is received.

Change-Id: I2f2dca02097d2781a2c324de1cd84074cc410cfe
Reviewed-on: http://map-gerrit.ric.broadcom.com/6428
Tested-by: Roman Bacik <rbacik@broadcom.com>
Tested-by: AutoBuild Status <jenkins_build_status_noreply@broadcom.com>
Reviewed-by: Abhinav Ratna <aratna@broadcom.com>
Reviewed-by: Srinath Mannam <srinath.mannam@broadcom.com>
Reviewed-by: Scott Branden <sbranden@broadcom.com>