SWMAPYVR-4518: Initial NS2 support
arm64: dts: use simple-bus for 32-bit peripherals

NS2 is a 64-bit processor. But a lot of its peripherals are 32-bit only.
To better represent the system architecture in dtsi, a "simple-bus" can
be deployed and have 32-bit peripherals defined underneath. The
simple-bus is mapped to the lower 4 GB (0x0000_0000 to 0xffff_ffff) of
the root bus

Change-Id: I2feaeb9a6b4cdb1b951957da30e1081370551a5a
Signed-off-by: Ray Jui <rjui@broadcom.com>