aio: add support for IOCB_CMD_POLL via aio thread helper
Applications that require a unified event loop occasionally have a need
to interface with libaries or other code that require notification on a
file descriptor becoming ready for read or write via poll.  Add support
for the aio poll operation to enable these use-cases by way of the thread
based aio helpers.

Signed-off-by: Benjamin LaHaise <ben.lahaise@solacesystems.com>
Signed-off-by: Benjamin LaHaise <bcrl@kvack.org>