PCI: iproc: Fix incorrect MSI address alignment
In the code to handle PAXB v2 based MSI steering, the logic aligns the
MSI register address to the size of supported inbound mapping range.
This is incorrect since it rounds "up" the starting address to the next
aligned address, but what we want is the starting address to be rounded
"down" to the aligned address.

This patch fixes the issue and allows MSI writes to be properly steered
to the GIC

Fixes: 4b073155fbd3 ("PCI: iproc: Add support for the next-gen PAXB
controller")

Signed-off-by: Ray Jui <ray.jui@broadcom.com>