vfio: Allow No-IOMMU mode without checking iommu_present()
Not allowing No-IOMMU mode for devices already having
iommu_ops on their bus is very conservative.

We now have IOMMU (such as ARM SMMU) which can bypass
transcations when IOMMU is not configured for a given
device. In addition, it is not necessary to have all
devices on bus to be upstream to an IOMMU on that bus.

Due above reasons, having iommu_present() check for
VFIO No-IOMMU mode is not appropriate.

Signed-off-by: Anup Patel <anup.patel@broadcom.com>